requirements.txt
azure-ai-language-questionanswering==1.1.0
azure-core==1.32.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
